network-monitor/requirements.txt
flask
paramiko
ping3
pyyaml
plotly